Fix chart core tests #685
Added fragment to provide dependency [1]
Fixed some Locale dependent tests


[1]
The chart core tests were failing due to an incompatible
java/osgi construct in org.apache.batik.transcoder.

The transcoder "bundle" is doing a Class.forname(String)
on a class that is not listed as a dependency.

This happens in class
org.apache.batik.transcoder.image.PNGTranscoder#getWriteAdapter
which is looking for
"org.apache.batik.ext.awt.image.codec.png.PNGTranscoderInternalCodecWriteAdapter"

To fool this "bundle", a fragment bundle is introduced with
o.a.b.transcoder as the master plugin that depends on the
"org.apache.batik.codec" bundle and in this way provides the dependency.